fix: Changed operator so that it does a server side apply rather than create for resources. (#470)

+++ crates/k8s-operator/src/services/bionic.rs
@@ -115,10 +115,14 @@ pub async fn deploy(client: Client, spec: BionicSpec, namespace: &str) -> Result
 pub async fn delete(client: Client, namespace: &str) -> Result<(), Error> {
     // Remove deployments
     let api: Api<Deployment> = Api::namespaced(client.clone(), namespace);
-    api.delete("bionic-gpt", &DeleteParams::default()).await?;
+    if api.get("bionic-gpt").await.is_ok() {
+        api.delete("bionic-gpt", &DeleteParams::default()).await?;
+    }
 
     // Remove services
     let api: Api<Service> = Api::namespaced(client, namespace);
-    api.delete("bionic-gpt", &DeleteParams::default()).await?;
+    if api.get("bionic-gpt").await.is_ok() {
+        api.delete("bionic-gpt", &DeleteParams::default()).await?;
+    }
     Ok(())
 }
